Add Dockerfile and CI
Add a Dockerfile to build an image for GoAuthing, which can be used to
run the app on a server without configuring systemd or other init
systems (e.g. Some commercial NAS devices may not support systemd).

Add two GitHub Actions jobs to build and push the Docker image to
GitHub Container Registry.